requirement.txt
alembic==1.13.1
beautifulsoup4==4.11.1
chardet==4.0.0
fake_useragent==1.4.0
fasttext==0.9.2
Flask==2.2.5
Flask_Mail==0.9.1
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
jieba==0.42.1
lxml==4.9.3
numpy==1.21.5
OpenCC==1.1.7
opencv_python==4.9.0.80
pandas==1.4.4
Pillow==10.3.0
py2neo==2021.2.4
pyltp==0.4.0
pypinyin==0.51.0
pytesseract==0.3.10
Requests==2.31.0
scikit_learn==1.3.0
selenium==4.19.0
SQLAlchemy==1.4.39
Werkzeug==2.2.3
WTForms==2.2.1
